CashMovementManagement/requirements.txt
asgiref==3.8.1
Django==5.2
django-admin-rangefilter==0.13.2
django-autocomplete-light==3.12.1
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8
packaging==24.2
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
wheel==0.45.1
